Refactor old effects into standalone boundscheck
