+++ tests/golden/test_effects/test_good_bound1.txt
@@ -1,6 +1,7 @@
-Reads:
-  { src : (8 * i + j) for (i,j) in Z if
-    0 <= i and i < (n + 7) / 8 and (n - 8 * i < 8 and (0 <= j and j < n - 8 * i)) }
-Writes:
-    { dst : (8 * i + j) for (i,j) in Z if
-    0 <= i and i < (n + 7) / 8 and (n - 8 * i < 8 and (0 <= j and j < n - 8 * i)) }
+def good_bound1(n: size, dst: R[n] @ DRAM, src: R[n] @ DRAM):
+    for i in seq(0, (n + 7) / 8):
+        if n - 8 * i >= 8:
+            pass
+        else:
+            for j in seq(0, n - 8 * i):
+                dst[8 * i + j] = src[8 * i + j]
