+++ tests/golden/test_effects/test_index3.txt
@@ -1,0 +1,4 @@
+def foo():
+    for i in seq(0, 0):
+        a: i8 @ DRAM
+        a = 0.0
