+++ tests/golden/test_effects/test_infereffects.txt
@@ -1,3 +1,2 @@
-Writes:
-    { A : (i,j) for (i,j) in Z if
-    0 <= i and i < 10 and (0 <= j and j < 20) }
+def bar(A: f32[10, 20] @ DRAM):
+    foo(A)
