+++ tests/golden/test_effects/test_mod1.txt
@@ -1,0 +1,2 @@
+def bar():
+    foo(10 % 3)
