+++ tests/golden/test_effects/test_race2.txt
@@ -1,6 +1,4 @@
-Reads:
-  { x : (i + 1,i) for (i) in Z if
-    0 <= i and i < n and i + 1 < n }
-Writes:
-    { x : (i,i) for (i) in Z if
-    0 <= i and i < n and i + 1 < n }
+def foo(n: size, x: R[n, n] @ DRAM):
+    for i in seq(0, n):
+        if i + 1 < n:
+            x[i, i] = x[i + 1, i]
