+++ tests/golden/test_effects/test_race3.txt
@@ -1,6 +1,5 @@
-Reads:
-  { x : (i + 1,i + 0) for (i) in Z if
-    0 <= i and i < n and i + 1 < n }
-Writes:
-    { x : (i,i) for (i) in Z if
-    0 <= i and i < n and i + 1 < n }
+def foo(n: size, x: R[n, n] @ DRAM):
+    y = x[1:n, 0:n]
+    for i in seq(0, n):
+        if i + 1 < n:
+            x[i, i] = y[i, i]
