+++ tests/golden/test_effects/test_seq_write1.txt
@@ -1,3 +1,4 @@
-Reads:
-  { A : (i) for (i) in Z if
-    0 <= i and i < n }
+def foo(n: size, A: i8[n] @ DRAM):
+    a: i8 @ DRAM
+    for i in seq(0, n):
+        a = A[i]
