+++ tests/golden/test_effects/test_stride_assert1.txt
@@ -1,0 +1,2 @@
+def bar(x: i8[30, 10] @ DRAM, y: i8[30, 16] @ GEMM_SCRATCH):
+    foo(30, 10, x, y)
