+++ tests/golden/test_effects/test_stride_assert11.txt
@@ -1,0 +1,4 @@
+def bar():
+    x: i8[8, 30, 10] @ DRAM
+    y: i8[50, 4, 100, 16] @ GEMM_SCRATCH
+    foo(30, 10, stride(x, 1), x[0, 0:30, 0:10], y[3, 1, 3:33, 0:16])
