+++ tests/test_config.py
@@ -371,120 +371,3 @@ def test_ld(golden):
     )
 
     assert f"{config_ld_i8}\n{ld_i8}" == golden
-
-
-"""
-
-    @proc
-    def ld_i8_v1_5(
-        n     : size,
-        m     : size,
-        scale : f32,
-        src   : [i8][n, m] @ DRAM,
-        dst   : [i8][n, 16] @ GEMM_SCRATCH,
-    ):
-        assert n <= 16
-        assert m <= 16
-        assert stride(src, 1) == 1
-        assert stride(dst, 0) == 16
-        assert stride(dst, 1) == 1
-        assert stride(src, 0) == ConfigLoad.src_stride
-
-        ConfigLoad.scale = scale
-        ConfigLoad.src_stride = stride(src, 0)
-
-        for i in seq(0, n):
-            for j in seq(0, m):
-                tmp : f32
-                tmp      = src[i,j]
-                tmp      = tmp * ConfigLoad.scale
-                dst[i,j] = tmp
-
-    @proc
-    def ld_i8_v2(
-        n     : size,
-        m     : size,
-        scale : f32,
-        src   : [i8][n, m] @ DRAM,
-        dst   : [i8][n, 16] @ GEMM_SCRATCH,
-    ):
-        assert n <= 16
-        assert m <= 16
-        assert stride(src, 1) == 1
-        assert stride(dst, 0) == 16
-        assert stride(dst, 1) == 1
-        assert stride(src, 0) == ConfigLoad.src_stride
-
-        config_ld_i8(scale, stride(src, 0))
-        do_ld_i8(n, m, src, dst)
-
-
-    new = (ld_i8.rename("new")
-                .replace(ld_i8, "for i in _:_")
-                .call_eqv(ld_i8_v2, "ld_i8(_, _, _, _, _)"))
-    print(new)
-
-
-
-
-def test_config_write3():
-    ConfigAB = new_config_f32()
-    @proc
-    def foo(n : size):
-        for i in seq(0, n):
-            ConfigAB.a = 3.0
-            ConfigAB.b = ConfigAB.a
-
-
-
-# This is fine
-def test_read_write2():
-    @proc
-    def foo(n : size, A : i8[n]):
-        a : i8
-        a = 4.0
-        for i in seq(0, n):
-            a    = 0.0
-            A[i] = a
-
-    foo.check_effects()
-
-
-
-def new_CONFIG_2f32():
-
-
-
-@proc
-def test_make_config(n: size, dst: R[n] @ DRAM, src: R[n] @ DRAM):
-    for i in seq(0, (n + 7) / 8):
-        if n - 8 * i >= 8:
-            pass
-        else:
-            for j in seq(0, n - 8 * i):
-                dst[8 * i + j] = src[8 * i + j]
-
-
-
-@instr("write_config({a}, {b})")
-def set_real_config( a : f32, b : f32):
-    REAL_CONFIG.a = a
-    REAL_CONFIG.b = b
-
-# st_config_ex
-@proc
-def set_config_a( a : f32 ):
-    DUMMY_CONFIG.a = a
-    set_real_config(DUMMY_CONFIG.a, DUMMY_CONFIG.b )
-
-# matmul_config_ex
-@proc
-def set_config_b( b : f32 ):
-    DUMMY_CONFIG.b = b
-    set_real_config(DUMMY_CONFIG.a, DUMMY_CONFIG.b )
-
-@instr("mvin...")
-def st_i8_v2( ..., stride : stride, ...)
-    set_config_a(stride)
-    do_st_i8(...)
-"""
